server/requirements.txt
aiXplain==0.2.11
asgiref==3.7.2
boto3==1.34.93
botocore==1.34.93
certifi==2024.2.2
charset-normalizer==3.3.2
click==7.1.2
cmdstanpy==1.2.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
filetype==1.2.0
fonttools==4.51.0
holidays==0.47
idna==3.7
importlib_resources==6.4.0
jmespath==1.0.1
kiwisolver==1.4.5
marshmallow==3.21.1
matplotlib==3.8.4
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
prophet==1.1.5
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
stanio==0.5.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.1